本の状態を判定するmodule名クラス名を変更したので呼び出し元も変更
